backend/requirements-dev.txt
# Production dependencies (from requirements.txt)
fastapi==0.104.1
uvicorn[standard]==0.24.0
authlib==1.3.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
python-multipart==0.0.6
strawberry-graphql[fastapi]==0.215.1
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
poppler-utils==0.1.0
pytesseract==0.3.10
httpx==0.25.2
apscheduler==3.10.4
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
watchdog==3.0.0
ipython==8.17.2
jupyter==1.0.0
httpie==3.2.2

# Performance monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Development tools
debugpy==1.8.0
ptvsd==4.0.2
memory-profiler==0.61.0
line-profiler==4.1.1

# Database development tools
pgcli==4.0.1
sqlparse==0.4.4

# API documentation
redoc==2.0.1

# Testing tools
factory-boy==3.3.0
faker==20.1.0
requests-mock==1.11.0

# Logging and debugging
structlog==23.2.0
rich==13.7.0
python-decouple==3.8

# Security scanning
bandit==1.7.5
safety==2.3.5

# Code quality
pylint==3.0.3
autopep8==2.0.4
yapf==0.40.2

# Development database
sqlite3